shoooweee a lot of movement
Got search somewhat working. Can properly match students to the search and get the proper students to display. issues still persist with using pagination after that.

Additionally, still need to handle when user clears the search box by either in reverting back to either page 1 or the page they were on, the latter would be more complicated but obviously feasible.